+++ CHANGELOG.md
@@ -4,12 +4,16 @@
 
 ### Added
 
-- Added simpler_command generator for Rails Applications
-- Added rspec:simpler_command generator for testing SimplerCommand in Rails
-
 ### Changed
 
 ### Removed
+
+## [0.1.1] - 2020-09-14
+
+### Added
+
+- Added simpler_command generator for Rails Applications
+- Added rspec:simpler_command generator for testing SimplerCommand in Rails
 
 ## [0.1.0] - 2020-09-11